add share button, factoring out query param methods to util
